+++ src/UploadQueue.h
@@ -88,8 +88,8 @@ private:
 #endif
 
 	std::set<CMD4Hash> suspendedUploadsSet;  // set for suspended uploads
-	uint32	m_nLastStartUpload;
-	uint32	m_lastSort;
+	uint64	m_nLastStartUpload;
+	uint64	m_lastSort;
 	bool	lastupslotHighID; // VQB lowID alternation
 	bool	m_allowKicking;
 	// This KnownFile collects all currently uploading clients for display in the upload list control
